False positive fixes
- Added new Speed check
- Fixing false positives with Fly (C) under blocks
- Fixed false positives with stepping on Fly (C) and Fly (D)
- Fixed liquid false positives with Fly (C) and (D)
- Fixed horizontal friction-based false positives, like switching between ice and regular blocks (used wrong friction)
- Fixed NoFall (B) false positive when in liquid and under block pressing space bar

+++ src/main/java/dev/brighten/ac/utils/world/EntityData.java
@@ -27,8 +27,8 @@ public class EntityData {
 
             //We cast this as a float since the fields are floats.
 
-            return new SimpleCollisionBox(new Vector(), fieldWidth.get(ventity),
-                    fieldLength.get(ventity));
+            return new SimpleCollisionBox(new Vector(), (float)fieldWidth.get(ventity),
+                    (float)fieldLength.get(ventity));
         }).copy();
     }
